Rename and restructure gem
