+++ Gemfile
@@ -2,7 +2,7 @@
 
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in resque-scheduler-unique-jobs.gemspec
+# Specify your gem's dependencies in resque-uniqueness.gemspec
 gemspec
 
 gem 'pry'
