+++ Rakefile
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-require 'resque_scheduler_unique_jobs'
+require 'resque/uniqueness'
 require 'bundler/gem_tasks'
 require 'rspec/core/rake_task'
 require 'resque/tasks'
