+++ lib/resque/uniqueness/version.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module Resque
+  module Uniqueness
+    VERSION = '0.1.0'
+  end
+end
