--- lib/resque_scheduler_unique_jobs/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module ResqueSchedulerUniqueJobs
-  VERSION = '0.1.0'
-end
