+++ resque-uniqueness.gemspec
@@ -2,27 +2,27 @@
 
 lib = File.expand_path('lib', __dir__)
 $LOAD_PATH.unshift(lib) unless $LOAD_PATH.include?(lib)
-require 'resque_scheduler_unique_jobs/version'
+require 'resque/uniqueness/version'
 
 Gem::Specification.new do |spec|
-  spec.name          = 'resque-scheduler-unique-jobs'
-  spec.version       = ResqueSchedulerUniqueJobs::VERSION
+  spec.name          = 'resque-uniqueness'
+  spec.version       = Resque::Uniqueness::VERSION
   spec.authors       = ['Olexandr Hoshylyk']
   spec.email         = ['gashuk95@gmail.com']
 
   spec.summary       = 'Gem for resque unique jobs'
   spec.description   = 'Implement unique jobs system like sidekiq-unique-jobs gem'
-  spec.homepage      = 'https://github.com/verbit/resque-scheduler-unique-jobs'
+  spec.homepage      = 'https://github.com/verbit/resque-uniqueness'
   spec.license       = 'MIT'
 
   # Prevent pushing this gem to RubyGems.org. To allow pushes either set the 'allowed_push_host'
   # to allow pushing to a single host or delete this section to allow pushing to any host.
   if spec.respond_to?(:metadata)
-    spec.metadata['allowed_push_host'] = 'https://github.com/verbit/resque-scheduler-unique-jobs'
+    spec.metadata['allowed_push_host'] = 'https://github.com/verbit/resque-uniqueness'
 
     spec.metadata['homepage_uri'] = spec.homepage
-    spec.metadata['source_code_uri'] = 'https://github.com/verbit/resque-scheduler-unique-jobs'
-    spec.metadata['changelog_uri'] = 'https://github.com/verbit/resque-scheduler-unique-jobs'
+    spec.metadata['source_code_uri'] = 'https://github.com/verbit/resque-uniqueness'
+    spec.metadata['changelog_uri'] = 'https://github.com/verbit/resque-uniqueness'
   else
     raise 'RubyGems 2.0 or newer is required to protect against ' \
       'public gem pushes.'
